feat: Add healthcheck endpoint
Add a new endpoint "/healthcheck" to the Flask application that returns a JSON response with the status "ok". This endpoint can be used for health monitoring purposes.

Refactor the __init__.py file in the app directory to import the "jsonify" function from the Flask module.

Update the whatsapp_quickstart.py file in the start directory to print the values of the ACCESS_TOKEN, RECIPIENT_WAID, PHONE_NUMBER_ID, and VERSION variables.

These changes address the need to add a healthcheck endpoint to the application and provide debugging information in the whatsapp_quickstart.py file.